\name{inner.function-package}
\Rdversion{1.1}
\alias{inner.function-package}
\docType{package}
\title{test package for inner functions}
\description{test package for inner functions...}
\details{
\tabular{ll}{Package: \tab inner.function\cr
Maintainer: \tab tdh\cr
Author: \tab tdh\cr
Version: \tab 1.0\cr
License: \tab GPL-3\cr
Title: \tab test package for inner functions\cr
Description: \tab test package for inner functions...\cr}
}
\author{tdh}
\references{
}
\keyword{ package }
\seealso{
}
\examples{}