\name{outer.function}
\Rdversion{1.1}
\alias{outer.function}
\title{outer function}
\description{this is a high-level package function that will be exported for all of your users.}
\usage{
outer.function(a, b)
}
\arguments{
  \item{a}{first arg}
  \item{b}{second arg}
}
\details{
}
\value{Return value of outer.function.}
\references{
}
\author{tdh}
\note{
}


\seealso{
}
\examples{}